fix: linter warning
