+++ src/hooks/useTheme.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { ThemeContext } from '@/app/theme/ThemeContext';
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) throw new Error('useTheme must be used within ThemeProvider');
+  return context;
+};
